requirements.txt
# development tools
ruff~=0.14
pre-commit~=3.8
ipykernel~=6.30
jupyterlab~=4.0
setuptools>=64.0.0

# job orchestration & utils
submitit~=1.5
tqdm~=4.67

# project packages
einops~=0.8
nibabel~=5.3
pynrrd~=1.1
matplotlib~=3.10
pydicom~=3.0
pandas>=2.1
scikit-learn~=1.7
numpy~=2.2
joblib~=1.5
PyYAML~=6.0
SimpleITK~=2.5
vtk==9.2.*
numba~=0.62
plotly~=6.5
networkx~=3.4
scikit-image~=0.25
pyvista~=0.46
seaborn~=0.13.2
torchio>=0.21.0
mrmr-selection~=0.2
neuroCombat~=0.2
xgboost~=2.1
